Add files via upload
This program explains the basics of main method and class.
When this code is compiled using java compiler(javac) the high level language code is converted to intermediate level(Byte Code) in form of a class file(helloworld.class).
And when we try to execute helloworld.class we use "java helloworld" this command line starts the JVM(also JRE) and JVM converts the IL code to ML code. And JVM also loads the executable image onto the RAM in JRE(A region which is allocated on the RAM by the OS for the purpose of execution of java program)and it is executed.